engine/client: expand `FetchResourceResponse` to include optional `headers`
This moves the existing resource-related types into a dedicated module (they probably already should have been!). It also revises their JSDoc to reflect some of the semantic expectations clarified in https://github.com/getodk/web-forms/issues/201.
