+++ packages/xforms-engine/src/instance/internal-api/InstanceConfig.ts
@@ -1,5 +1,5 @@
-import type { FetchFormAttachment, FetchResource } from '../../client/EngineConfig.ts';
 import type { OpaqueReactiveObjectFactory } from '../../client/OpaqueReactiveObjectFactory.ts';
+import type { FetchFormAttachment, FetchResource } from '../../client/resources.ts';
 import type { CreateUniqueId } from '../../lib/unique-id.ts';
 
 export interface InstanceConfig {
